requirements.txt
inky
python-dotenv
requests
Selenium-Screenshot
selenium
git+https://www.github.com/hbldh/hitherdither
